Made changes to allow for pytest to run properly, commented out all test groups besides user and student tests to begin. Ensured that unit and integration tests can be ran separate from eachother.
