+++ App/tests/__init__.py
@@ -1,12 +1,12 @@
 from .test_app import *
-from .test_student import *
-from .test_staff import *
-from .test_review import *
-from .test_recommendation import *
-from .test_incident_report import *
-from .test_accomplishment import *
-from .test_grades import *
-from .test_admin import *
-from .test_karma import *
-from .test_nltk import *
+#from .test_student import *
+# from .test_staff import *
+# from .test_review import *
+# from .test_recommendation import *
+# from .test_incident_report import *
+# from .test_accomplishment import *
+# from .test_grades import *
+# from .test_admin import *
+# from .test_karma import *
+# from .test_nltk import *
