+++ App/tests/locustfile.py
@@ -1,9 +1,9 @@
-from locust import HttpUser, task, between
+# from locust import HttpUser, task, between
 
-class WebsiteUser(HttpUser):
-    wait_time = between(5, 9)  # Time between consecutive requests
+# class WebsiteUser(HttpUser):
+#     wait_time = between(5, 9)  # Time between consecutive requests
 
-    @task
-    def index_page(self):
-        self.client.get("/")  # Make a GET request to the root URL
+#     @task
+#     def index_page(self):
+#         self.client.get("/")  # Make a GET request to the root URL
