+++ App/tests/test_nltk.py
@@ -1,20 +1,20 @@
-import os, tempfile, pytest, logging, unittest
+# import os, tempfile, pytest, logging, unittest
 
-from App.main import create_app
-from App.database import db, create_db
-from App.controllers import (
-    analyze_sentiment,
-    analyze_text
-)
+# from App.main import create_app
+# from App.database import db, create_db
+# from App.controllers import (
+#     analyze_sentiment,
+#     analyze_text
+# )
 
-#class NLTKUnitTests(unittest.TestCase):
+# #class NLTKUnitTests(unittest.TestCase):
 
-    #def test_analyze_text(self):
-        #processed_text = analyze_text("This is amazing")
-        #assert processed_text is not None
+#     #def test_analyze_text(self):
+#         #processed_text = analyze_text("This is amazing")
+#         #assert processed_text is not None
 
-    #def test_analyze_sentiment(self):
-      #scores = analyze_sentiment("This is amazing")
-      #assert scores is not None
-      #if scores is not None:
-          #assert scores == 10.0
+#     #def test_analyze_sentiment(self):
+#       #scores = analyze_sentiment("This is amazing")
+#       #assert scores is not None
+#       #if scores is not None:
+#           #assert scores == 10.0
